pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "tgram-analytics-server"
version = "0.1.0"
description = "Self-hosted, privacy-first analytics platform controlled via Telegram"
readme = "README.md"
license = { text = "FSL-1.1-ALv2" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.3.0",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "ua-parser>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "fakeredis>=2.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "E501",   # line too long — handled by formatter
    "B008",   # Depends() in argument defaults is the FastAPI pattern
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
